requirements.txt
dpath>=1.4.0
pyDes>=2.0.1
ldap3>=2.3
python3-ldap>=0.9.8.4
dnspython>=1.15.0